This is the first try at committing some of the code.  It is very preliminary (of course) and most likely using methods for accessing the YBK files that need to be changed in order to properly handle embedded graphics (i.e., img tags) and links to other "chapters" or internal files within the YBK file.  I have already started the work of converting file access to use a ContentProvider instead but it is not actually in use.  The code is able to read any files in the default library directory (/sdcard/) and produce a list of files as well as bringing up the index page of each YBK file when you touch the name of the YanCEyBook in the list.
The reason I am converting to a ContentProvider (YbkProvider) is that currently the loadData() method is used to load content into the WebView that is used for showing the text of the chapters.  From the reading I have done, loadData() does not loading new text from inside a YBK file when clicking (touching?) on an anchor tag.  Instead, loadUrl() must be used so we must use a ContentProvider so content:// style URI's can be used.