Converted to a provider and database implementation.  Still only allows for traversal of a 2 deep category search.  I have a major issue with titles that only have one or no categories to resolve.
Also, this temporarily uses an xml file in the raw resources to initially populate the database.  It's slow on the first load only.  I'm still trying to work out how to create a good database and them move it to the assets for the project and use it instead of building it from xml.