Issue 27: Added bookmark capability.
Fixed a bug in the schema design of the history table.  The bookmark_id field was defined as INTEGER NULL UNIQUE.  It was possible to have multiple records with null as a value in that field.  This appeared to cause sorting problems.  I don't suppose that SQLite should have allowed it. I have changed bookmark_id to INTEGER NULL.

WARNING: This updates the version of the database to 12 and will therefore cause the regeneration of the database.  Users installing the next update should be made aware of this.

I have also made a few miscellaneous code cleanups in order to reduce the size of the compilation units (i.e., class files).